user input of accesskey works. Need to check if it works in the container

+++ src/main/java/com/fec/restclient/service/FileWriterService.java
@@ -6,6 +6,7 @@ import java.io.*;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.ArrayList;
+import java.util.List;
 
 @Component
 public class FileWriterService {
@@ -14,6 +15,7 @@ public class FileWriterService {
     String lineValue;
     PrintWriter writer;
     String awsAccessLine = "awsAccessKey";
+    String fileName;
 
     public boolean createFile(String filename){
 
@@ -22,6 +24,7 @@ public class FileWriterService {
         // if file doesn't exist, create it
         if(!tmpDir.exists()) {
             try {
+                System.out.println("New file location "+filename);
                 this.writer = new PrintWriter(filename, "UTF-8");
                 return true;
             } catch (FileNotFoundException e) {
@@ -53,16 +56,16 @@ public class FileWriterService {
 
         ArrayList keyList = new ArrayList();
 
+
         try {
             Files.lines(Paths.get(fileName))
                     .map(s -> s.trim())
-                    .filter(s -> s.startsWith("awsAccessKey"))
+                    .filter(s -> s.startsWith(keyName))
                     .forEach(keyList::add);
         } catch (IOException e) {
             e.printStackTrace();
         }
-
-        System.out.println(keyList.get(0).toString());
+        System.out.println("WHERE IS THIS?");
         return keyList.get(0).toString();
 
     }
@@ -108,10 +111,47 @@ public class FileWriterService {
 
     }
 
+    public List<String> fileToArray(){
+        List<String> result = null;
+
+        System.out.println("filename set: "+ this.fileName);
+        File file = new File(this.fileName);
+
+        if(file.exists()) {
+            System.out.println("File exists...");
+            try {
+                result = Files.readAllLines(Paths.get(this.fileName));
+                if(result.size()== 0){
+                    System.out.println("size was zero deleting...");
+                    this.deleteFile(this.fileName);
+                    return null;
+                }
+
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+        }
+        else{
+            System.out.println("File does not exist...");
+        }
+        return result;
+    }
+
     public void deleteFile(String fileName){
 
         File tmpDir = new File(fileName);
-        tmpDir.delete();
+        if (tmpDir.exists()){
+            System.out.println("exists...");
+            tmpDir.delete();
+        }
+        else{
+            System.out.println("Doesn't exist");
+        }
+        System.out.println("File deleted...");
     }
 
+
+    public void setFileName(String fileName) {
+        this.fileName = fileName;
+    }
 }
